move PromptForNickname
