+++ src/holofuel/components/wrappers/PromptForNickname/PromptForNickname.js
