+++ src/holofuel/components/wrappers/PromptForNickname/index.js
@@ -1,0 +1,3 @@
+import Component from './PromptForNickname'
+
+export default Component
